pyproject.toml
[tool.poetry]
name = "mcp-implementation"
version = "0.1.0"
description = "Análisis de Incidentes en Medios Sociales usando MCP"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
pydantic = "^2.5.0"
fastapi = "^0.104.0"
uvicorn = "^0.24.0"
tweepy = "^4.14.0"
spacy = "^3.7.2"
sqlalchemy = "^2.0.23"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.0"
plotly = "^5.18.0"
jinja2 = "^3.1.2"
pandas = "^2.1.3"
apscheduler = "^3.10.4"

[tool.poetry.dev-dependencies]
pytest = "^7.4.0"
black = "^23.10.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api" 